move registry papckage to vttablet
