+++ go/vt/vttablet/registry/registry_test.go
